move Today-related hooks into a subfolder
