+++ client/src/components/Today/hooks/use-detailed-activity-modal.ts
